+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Feature
 - Change venue when match warmup starts
 - New endpoint `/tournament/<id>/matches/result` returning results for all matches in tournament
+- Added more general statistics for each tournament to `/tournament/<id>/statistics`
 
 #### Changed
 - Switched from string to time for datetimes to correctly support timezones